stepderinitions main class refactor
